Refactoring part2
