+++ packages/mobile/src/features/auth/action/handleRegistrationAlert.tsx
