+++ packages/mobile/src/features/auth/page/SingUpScrenn.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import styled from 'styled-components/native';
 import {View, Text, ScrollView} from 'react-native';
 import Header from '../../common/components/Header';
-import InputSection from '../components/InputSection';
+import InputSection from '../components/SingUp/InputSection';
 import Footer from '../components/Footer';
 
 const Wrapper = styled.View`
